feat(useAPIForVADCDataDictionary): Separated out Container and Loading components, reorganized code into subfolder
