+++ src/Analysis/AtlasDataDictionary/AtlasDataDictionaryTable/AtlasDataDictionaryTable.tsx
@@ -1,25 +1,25 @@
 import React, { useState } from 'react';
 import { Table } from '@mantine/core';
-import ColumnHeaders from '../Components/ColumnHeaders/ColumnHeaders';
-import EntriesHeader from '../Components/EntriesHeader/EntriesHeader';
-import SearchBar from '../Components/SearchBar/SearchBar';
-import TableRow from '../Components/TableRow/TableRow';
-import PaginationControls from '../Components/PaginationControls/PaginationControls';
-import { ISortConfig } from '../Interfaces/Interfaces';
-import PreprocessTableData from '../Utils/PreprocessTableData';
-import InitialDataDictionaryTableState from '../Utils/InitialDataDictionaryTableState';
+import ColumnHeaders from './Components/ColumnHeaders/ColumnHeaders';
+import EntriesHeader from './Components/EntriesHeader/EntriesHeader';
+import SearchBar from './Components/SearchBar/SearchBar';
+import TableRow from './Components/TableRow/TableRow';
+import PaginationControls from './Components/PaginationControls/PaginationControls';
+import { ISortConfig } from './Interfaces/Interfaces';
+import PreprocessTableData from './Utils/PreprocessTableData';
+import InitialDataDictionaryTableState from './Utils/InitialDataDictionaryTableState';
 import {
   DetermineNextSortDirection,
   SortDataWithDirection,
-} from '../Utils/SortUtils';
-import ManageColumns from '../Components/ManageColumns/ManageColumns';
-import DefaultAtlasColumnManagement from '../Utils/DefaultAtlasColumnManagement';
+} from './Utils/SortUtils';
+import ManageColumns from './Components/ManageColumns/ManageColumns';
+import DefaultAtlasColumnManagement from './Utils/DefaultAtlasColumnManagement';
 
 const AtlasDataDictionaryTable = ({ TableData }) => {
   const preprocessedTableData = PreprocessTableData(TableData);
   const [data, setData] = useState(preprocessedTableData);
   const [dataDictionaryTableState, setDataDictionaryTableState] = useState(
-    InitialDataDictionaryTableState
+    InitialDataDictionaryTableState,
   );
 
   const {
@@ -32,19 +32,17 @@ const AtlasDataDictionaryTable = ({ TableData }) => {
     columnManagement,
   } = dataDictionaryTableState;
 
-  const entriesHeaderStart =
-    dataDictionaryTableState.entriesShown *
-      dataDictionaryTableState.currentPage -
-    dataDictionaryTableState.entriesShown +
-    1;
+  const entriesHeaderStart = dataDictionaryTableState.entriesShown
+      * dataDictionaryTableState.currentPage
+    - dataDictionaryTableState.entriesShown
+    + 1;
 
-  const entriesHeaderStop =
-    dataDictionaryTableState.entriesShown *
-    dataDictionaryTableState.currentPage;
+  const entriesHeaderStop = dataDictionaryTableState.entriesShown
+    * dataDictionaryTableState.currentPage;
 
   const paginatedData = data.slice(
     entriesShown * currentPage - entriesShown,
-    entriesShown * currentPage
+    entriesShown * currentPage,
   );
 
   const handleTableChange = (
@@ -57,7 +55,7 @@ const AtlasDataDictionaryTable = ({ TableData }) => {
       | 'sortConfig'
       | 'columnManagementUpdateOne'
       | 'columnManagementReset',
-    eventData: any
+    eventData: any,
   ) => {
     if (event === 'openDropdown') {
       setDataDictionaryTableState({
@@ -68,7 +66,7 @@ const AtlasDataDictionaryTable = ({ TableData }) => {
       setDataDictionaryTableState({
         ...dataDictionaryTableState,
         openDropdowns: openDropdowns.filter(
-          (dropdownNumber: number) => dropdownNumber !== eventData
+          (dropdownNumber: number) => dropdownNumber !== eventData,
         ),
       });
     } else if (event === 'currentPage') {
@@ -90,9 +88,9 @@ const AtlasDataDictionaryTable = ({ TableData }) => {
         columnManagement: DefaultAtlasColumnManagement,
       });
     } else if (
-      event === 'entriesShown' ||
-      event === 'searchTerm' ||
-      event === 'sortConfig'
+      event === 'entriesShown'
+      || event === 'searchTerm'
+      || event === 'sortConfig'
     ) {
       setDataDictionaryTableState({
         ...dataDictionaryTableState,
@@ -101,7 +99,7 @@ const AtlasDataDictionaryTable = ({ TableData }) => {
       });
     } else {
       throw new Error(
-        `handleTableChange called with invalid parameters: event: ${event}, eventData: ${eventData}`
+        `handleTableChange called with invalid parameters: event: ${event}, eventData: ${eventData}`,
       );
     }
   };
@@ -109,7 +107,7 @@ const AtlasDataDictionaryTable = ({ TableData }) => {
   const handleSort = (sortKey: string) => {
     const newDirection: ISortConfig['direction'] = DetermineNextSortDirection(
       sortConfig as ISortConfig,
-      sortKey
+      sortKey,
     );
     const sortedData = SortDataWithDirection(data, newDirection, sortKey);
     // if column is set to off reset to initial sort
