+++ src/Analysis/AtlasDataDictionary/AtlasDataDictionaryTable/Components/Icons/CollapseIcon.tsx
